+++ CHANGELOG.md
@@ -6,6 +6,22 @@ rename `Unreleased` topic with the new version tag. Finally, create a new `Unrel
 
 ## Unreleased
 
+### Core: events
+- New window events: `windowMinimize`, `windowRestore`, `windowMaximize`, `windowFullScreenEnter`, and `windowFullScreenExit`
+
+## v6.4.0
+
+### API: storage
+- Implement `storage.clear()` and `storage.removeData(key)` functions to remove saved storage data.
+
+### Improvements/bugfixes
+- Fix the draggable region not working issue on Windows.
+- Replace deprecated macOS API with suitable modern APIs in the codebase.
+- Static file server enhancements.
+- Improve `window.getPositon()` in macOS.
+- Save the correct window size and position when a maximized/minimized window is being closed on Windows.
+- Fix the window disappearing issue while restoring the window on Windows
+
 ## v6.3.0
 
 ### Single-executable mode